As more settings have been added, I extracted the PluginTask from the Plugin to make it easier to understand when reading the code.
