+++ src/main/scala/org/embulk/output/s3_parquet/S3ParquetOutputPlugin.scala
@@ -1,31 +1,21 @@
 package org.embulk.output.s3_parquet
 
 import java.nio.file.{Files, Paths}
-import java.util.{
-  IllegalFormatException,
-  Locale,
-  Optional,
-  List => JList,
-  Map => JMap
-}
+import java.util.{IllegalFormatException, Locale, List => JList}
 
 import com.amazonaws.services.s3.model.CannedAccessControlList
 import org.apache.parquet.column.ParquetProperties
 import org.apache.parquet.hadoop.ParquetWriter
 import org.apache.parquet.hadoop.metadata.CompressionCodecName
 import org.embulk.config.{
-  Config,
-  ConfigDefault,
   ConfigDiff,
   ConfigException,
   ConfigSource,
-  Task,
   TaskReport,
   TaskSource
 }
-import org.embulk.output.s3_parquet.S3ParquetOutputPlugin.{
+import org.embulk.output.s3_parquet.PluginTask.{
   ColumnOptionTask,
-  PluginTask,
   TypeOptionTask
 }
 import org.embulk.output.s3_parquet.aws.Aws
@@ -41,92 +31,10 @@ import org.embulk.spi.{
   TransactionalPageOutput
 }
 import org.embulk.spi.time.TimestampFormatter
-import org.embulk.spi.time.TimestampFormatter.TimestampColumnOption
 import org.embulk.spi.util.Timestamps
 import org.slf4j.{Logger, LoggerFactory}
 
 import scala.util.chaining._
-
-object S3ParquetOutputPlugin {
-
-  trait PluginTask extends Task with TimestampFormatter.Task with Aws.Task {
-
-    @Config("bucket")
-    def getBucket: String
-
-    @Config("path_prefix")
-    @ConfigDefault("\"\"")
-    def getPathPrefix: String
-
-    @Config("sequence_format")
-    @ConfigDefault("\"%03d.%02d.\"")
-    def getSequenceFormat: String
-
-    @Config("file_ext")
-    @ConfigDefault("\"parquet\"")
-    def getFileExt: String
-
-    @Config("compression_codec")
-    @ConfigDefault("\"uncompressed\"")
-    def getCompressionCodecString: String
-
-    def getCompressionCodec: CompressionCodecName
-    def setCompressionCodec(v: CompressionCodecName): Unit
-
-    @Config("column_options")
-    @ConfigDefault("{}")
-    def getColumnOptions: JMap[String, ColumnOptionTask]
-
-    @Config("canned_acl")
-    @ConfigDefault("\"private\"")
-    def getCannedAclString: String
-
-    def setCannedAcl(v: CannedAccessControlList): Unit
-
-    def getCannedAcl: CannedAccessControlList
-
-    @Config("block_size")
-    @ConfigDefault("null")
-    def getBlockSize: Optional[Int]
-
-    @Config("page_size")
-    @ConfigDefault("null")
-    def getPageSize: Optional[Int]
-
-    @Config("max_padding_size")
-    @ConfigDefault("null")
-    def getMaxPaddingSize: Optional[Int]
-
-    @Config("enable_dictionary_encoding")
-    @ConfigDefault("null")
-    def getEnableDictionaryEncoding: Optional[Boolean]
-
-    @Config("buffer_dir")
-    @ConfigDefault("null")
-    def getBufferDir: Optional[String]
-
-    @Config("catalog")
-    @ConfigDefault("null")
-    def getCatalog: Optional[CatalogRegistrator.Task]
-
-    @Config("type_options")
-    @ConfigDefault("{}")
-    def getTypeOptions: JMap[String, TypeOptionTask]
-  }
-
-  trait ColumnOptionTask
-      extends Task
-      with TimestampColumnOption
-      with LogicalTypeOption
-
-  trait TypeOptionTask extends Task with LogicalTypeOption
-
-  trait LogicalTypeOption {
-
-    @Config("logical_type")
-    def getLogicalType: Optional[String]
-  }
-}
 
 class S3ParquetOutputPlugin extends OutputPlugin {
